+++ pyrpckit/codegen/__init__.py
@@ -9,7 +9,7 @@ from pyrpckit.codegen.ir import ClientIr, UnsupportedSchemaError, build_ir
 from pyrpckit.codegen.options import PythonClientOptions, TypeScriptClientOptions
 from pyrpckit.codegen.writer import MANIFEST, write_files
 
-LAYOUT_VERSION = 7
+LAYOUT_VERSION = 8
 
 __all__ = [
     "ClientIr",